Refactor pygame controller into modules
